grandma ask APIのビジネスロジックをlib/grandma/モジュールへ分離する
- lib/grandma/types.ts: 共有型定義
- lib/grandma/consultUtils.ts: 純粋ユーティリティ関数（地理計算・キーワード抽出・エラーレスポンス構築など）
- lib/grandma/vendorSearch.ts: Supabaseデータ取得（N+1クエリをPromise.all並列化で改善）
- lib/grandma/promptBuilder.ts: LLMプロンプト構築・ストリーミング出力パース
- lib/grandma/abuseDetection.ts: 不正検知ロジック
- route.ts: HTTPレイヤーのみ残す（1820行→906行）
- unit tests: consultUtils・promptBuilder の主要パスをカバー（48テスト）
- SHOP_IDS空文字パースのバグを修正（0が混入していた問題）
